+++ ackpine-core/src/main/kotlin/ru/solrudev/ackpine/impl/installer/InstallSessionFactory.kt
@@ -20,6 +20,7 @@ package ru.solrudev.ackpine.impl.installer
 
 import android.annotation.SuppressLint
 import android.content.Context
+import android.os.Build
 import android.os.Handler
 import androidx.annotation.RestrictTo
 import androidx.annotation.WorkerThread
@@ -39,13 +40,13 @@ import ru.solrudev.ackpine.impl.installer.session.IntentBasedInstallSession
 import ru.solrudev.ackpine.impl.installer.session.SessionBasedInstallSession
 import ru.solrudev.ackpine.impl.installer.session.getSessionBasedSessionCommitProgressValue
 import ru.solrudev.ackpine.impl.installer.session.helpers.PROGRESS_MAX
+import ru.solrudev.ackpine.impl.session.CompletableProgressSession
 import ru.solrudev.ackpine.installer.InstallFailure
 import ru.solrudev.ackpine.installer.parameters.InstallParameters
 import ru.solrudev.ackpine.installer.parameters.InstallerType
 import ru.solrudev.ackpine.installer.parameters.PackageSource
 import ru.solrudev.ackpine.resources.ResolvableString
 import ru.solrudev.ackpine.session.Progress
-import ru.solrudev.ackpine.session.ProgressSession
 import ru.solrudev.ackpine.session.Session
 import ru.solrudev.ackpine.session.Session.State.Committed
 import ru.solrudev.ackpine.session.Session.State.Succeeded
@@ -62,13 +63,13 @@ internal interface InstallSessionFactory {
 		id: UUID,
 		notificationId: Int,
 		dbWriteSemaphore: BinarySemaphore
-	): ProgressSession<InstallFailure>
+	): CompletableProgressSession<InstallFailure>
 
 	@WorkerThread
 	fun create(
 		session: SessionEntity.InstallSession,
 		completeIfSucceeded: Boolean = false
-	): ProgressSession<InstallFailure>
+	): CompletableProgressSession<InstallFailure>
 
 	fun resolveNotificationData(notificationData: NotificationData, name: String): NotificationData
 }
@@ -93,7 +94,7 @@ internal class InstallSessionFactoryImpl internal constructor(
 		id: UUID,
 		notificationId: Int,
 		dbWriteSemaphore: BinarySemaphore
-	): ProgressSession<InstallFailure> = when (parameters.installerType) {
+	): CompletableProgressSession<InstallFailure> = when (parameters.installerType) {
 		InstallerType.INTENT_BASED -> IntentBasedInstallSession(
 			applicationContext,
 			apk = parameters.apks.toList().singleOrNull() ?: throw SplitPackagesNotSupportedException(),
@@ -132,7 +133,7 @@ internal class InstallSessionFactoryImpl internal constructor(
 	override fun create(
 		session: SessionEntity.InstallSession,
 		completeIfSucceeded: Boolean
-	): ProgressSession<InstallFailure> = when (session.installerType) {
+	): CompletableProgressSession<InstallFailure> = when (session.installerType) {
 		InstallerType.INTENT_BASED -> createIntentBasedInstallSession(session, completeIfSucceeded)
 		InstallerType.SESSION_BASED -> createSessionBasedInstallSession(session, completeIfSucceeded)
 	}
@@ -224,8 +225,13 @@ internal class InstallSessionFactoryImpl internal constructor(
 			return session
 		}
 		val progressThreshold = (applicationContext.getSessionBasedSessionCommitProgressValue() * PROGRESS_MAX).toInt()
-		if (initialProgress.progress >= progressThreshold) {
-			// means that actual installation is ongoing or is completed
+		val isInstallationOngoingOrCompleted = initialProgress.progress >= progressThreshold
+		// We can't rely on session's progress on API 31-32 with requireUserAction == false, so we don't block clients
+		// from committing on these versions if user's confirmation was already launched previously.
+		val shouldBlockCommitIfIsOngoing = installSession.session.requireUserAction
+				|| Build.VERSION.SDK_INT !in 31..32
+				|| installSession.wasConfirmationLaunched != true
+		if (shouldBlockCommitIfIsOngoing && isInstallationOngoingOrCompleted) {
 			session.notifyCommitted() // block clients from committing
 		}
 		// If app is killed while installing but system installer activity remains visible,