+++ ackpine-core/src/main/kotlin/ru/solrudev/ackpine/installer/parameters/InstallParameters.kt
@@ -21,6 +21,7 @@ import android.content.pm.PackageInstaller
 import android.net.Uri
 import android.os.Build
 import androidx.annotation.RequiresApi
+import ru.solrudev.ackpine.DelicateAckpineApi
 import ru.solrudev.ackpine.exceptions.SplitPackagesNotSupportedException
 import ru.solrudev.ackpine.session.parameters.Confirmation
 import ru.solrudev.ackpine.session.parameters.ConfirmationAware
@@ -364,7 +365,12 @@ public class InstallParameters private constructor(
 
 		/**
 		 * Sets [InstallParameters.requireUserAction].
+		 *
+		 * This is a **delicate** API. This option is unstable for use on different Android versions from different
+		 * vendors. It's recommended to avoid using it on API level < 33 and on devices with modified OS package
+		 * installer, most notably from Chinese vendors, unless your app is privileged for silent installs.
 		 */
+		@DelicateAckpineApi
 		public fun setRequireUserAction(requireUserAction: Boolean): Builder = apply {
 			this.requireUserAction = requireUserAction
 		}