+++ build-logic/src/main/kotlin/ru/solrudev/ackpine/gradle/AckpineCommonExtension.kt
@@ -18,6 +18,8 @@ package ru.solrudev.ackpine.gradle
 
 import com.android.build.api.dsl.CommonExtension
 
+internal typealias IdListener = (id: String) -> Unit
+
 /**
  * Common extension properties for Ackpine `library` and `sample` plugins.
  */
@@ -33,7 +35,6 @@ public abstract class AckpineCommonExtension(
 	 * Ackpine library ID used in namespace of the generated R and BuildConfig classes, in artifact ID
 	 * and in application ID.
 	 */
-	@Suppress("MemberVisibilityCanBePrivate")
 	public open var id: String
 		get() = _id
 		set(value) {