+++ build-logic/src/main/kotlin/ru/solrudev/ackpine/gradle/Constants.kt
@@ -20,15 +20,15 @@ import org.gradle.api.JavaVersion
 
 public object Constants {
 	public const val PACKAGE_NAME: String = "ru.solrudev.ackpine"
-	public const val JDK_VERSION: Int = 17
+	public const val JDK_VERSION: Int = 21
 	public const val MIN_SDK: Int = 16
 	public const val COMPILE_SDK: Int = 35
-	public const val BUILD_TOOLS_VERSION: String = "35.0.0"
+	public const val BUILD_TOOLS_VERSION: String = "35.0.1"
 }
 
 public object SampleConstants {
 	public const val PACKAGE_NAME: String = "ru.solrudev.ackpine.sample"
 	public const val MIN_SDK: Int = 21
 	public const val TARGET_SDK: Int = 35
-	public val JAVA_VERSION: JavaVersion = JavaVersion.VERSION_17
+	public val JAVA_VERSION: JavaVersion = JavaVersion.VERSION_21
 }