+++ build-logic/src/main/kotlin/ru/solrudev/ackpine/gradle/documentation/DokkaConventionPlugin.kt
@@ -18,13 +18,16 @@ package ru.solrudev.ackpine.gradle.documentation
 
 import org.gradle.api.Plugin
 import org.gradle.api.Project
+import org.gradle.api.provider.Provider
 import org.gradle.kotlin.dsl.apply
 import org.gradle.kotlin.dsl.assign
 import org.gradle.kotlin.dsl.configure
+import org.gradle.kotlin.dsl.findByType
 import org.gradle.kotlin.dsl.named
 import org.jetbrains.dokka.gradle.DokkaExtension
 import org.jetbrains.dokka.gradle.DokkaPlugin
 import org.jetbrains.dokka.gradle.engine.plugins.DokkaHtmlPluginParameters
+import ru.solrudev.ackpine.gradle.AckpineLibraryExtension
 import ru.solrudev.ackpine.gradle.versioning.versionNumber
 import java.net.URI
 
@@ -32,10 +35,15 @@ public class DokkaConventionPlugin : Plugin<Project> {
 
 	override fun apply(target: Project): Unit = target.run {
 		pluginManager.apply(DokkaPlugin::class)
-		configureDokka()
+		val internalPackages = extensions
+			.findByType<AckpineLibraryExtension>()
+			?.internalPackages ?: provider { emptySet() }
+		configureDokka(internalPackages)
 	}
 
-	private fun Project.configureDokka() = extensions.configure<DokkaExtension> {
+	private fun Project.configureDokka(
+		internalPackages: Provider<Set<String>>
+	) = extensions.configure<DokkaExtension> {
 		moduleVersion = versionNumber.get().toString()
 		pluginsConfiguration.named<DokkaHtmlPluginParameters>("html") {
 			customAssets.from(isolated.rootProject.projectDirectory.file("docs/images/logo-icon.svg"))
@@ -49,6 +57,16 @@ public class DokkaConventionPlugin : Plugin<Project> {
 			externalDocumentationLinks.register("guava") {
 				url = URI("https://guava.dev/releases/snapshot/api/docs/")
 				packageListUrl = URI("https://guava.dev/releases/snapshot/api/docs/element-list")
+			}
+			val suppressedPackages = internalPackages.get()
+			if (suppressedPackages.isNotEmpty()) {
+				val internalSources = sourceRoots.asFileTree.matching {
+					for (packageName in suppressedPackages) {
+						val packagePath = packageName.replace('.', '/')
+						include("**/$packagePath/**")
+					}
+				}
+				suppressedFiles.from(internalSources)
 			}
 		}
 	}